refactoring InfinispanServerTestResource
