+++ data-index/data-index-service/src/test/java/org/kie/kogito/index/DataIndexInfinispanServerTestResource.java
@@ -1,0 +1,14 @@
+package org.kie.kogito.index;
+
+import org.kie.kogito.persistence.infinispan.InfinispanServerTestResource;
+
+public class DataIndexInfinispanServerTestResource extends InfinispanServerTestResource {
+
+    @Override
+    public boolean shouldCleanCache(String cacheName) {
+        return cacheName.equals("processinstances")
+                || cacheName.endsWith("_domain")
+                || cacheName.equals("jobs")
+                || cacheName.equals("usertaskinstances");
+    }
+}
