+++ persistence-commons/persistence-commons-infinispan/src/test/java/org/kie/kogito/persistence/infinispan/QuarkusTestResourceWithCleanupLifecycleManager.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.index;
+package org.kie.kogito.persistence.infinispan;
 
 import java.util.Arrays;
